+++ db-data/one-off-scripts/2022-02-11-cipher-unlock-to-lock.ts
@@ -1,11 +1,11 @@
 // Execute this file with:
 // npx ts-node db-data/2022-02-11-cipher-unlock-to-lock.ts
 
-import './_env'
+import '../_env'
 
 import bluebird from 'bluebird'
 
-import { firebase } from '../pages/api/_services'
+import { firebase } from '../../pages/api/_services'
 import electionVotes from './elections-votes.json'
 
 let found = 0
