+++ db-data/one-off-scripts/2023-05-02-analyze-encryption-details.ts
@@ -1,4 +1,4 @@
-import { RP, pointToString, stringToPoint } from '../src/crypto/curve'
+import { RP, pointToString, stringToPoint } from '../../src/crypto/curve'
 
 const encodeds = ['']
