+++ db-data/one-off-scripts/2023-05-02-count-encryption-holes.ts
@@ -1,15 +1,15 @@
-import './_env'
+import '../_env'
 
 import { inspect } from 'util'
 
 import { keyBy, mapValues, pick } from 'lodash'
 import UAParser from 'ua-parser-js'
 
-import { firebase } from '../pages/api/_services'
-import { RP, pointToString } from '../src/crypto/curve'
-import decrypt from '../src/crypto/decrypt'
-import { CipherStrings } from '../src/crypto/stringify-shuffle'
-import { tallyVotes } from '../src/status/tally-votes'
+import { firebase } from '../../pages/api/_services'
+import { RP, pointToString } from '../../src/crypto/curve'
+import decrypt from '../../src/crypto/decrypt'
+import { CipherStrings } from '../../src/crypto/stringify-shuffle'
+import { tallyVotes } from '../../src/status/tally-votes'
 
 const election_id = '1680323766282'
